% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apartments.R
\docType{data}
\name{apartments}
\alias{apartments}
\alias{apartments_test}
\alias{apartmentsTest}
\title{Apartments Data}
\format{a data frame with 1000 rows and 6 columns}
\usage{
data(apartments)
}
\description{
Datasets \code{apartments} and \code{apartmentsTest} are artificial,
generated form the same model.
Structure of the dataset is copied from real dataset from \code{PBImisc} package,
but they were generated in a way to mimic effect of Anscombe quartet for complex black box models.
}
\details{
\itemize{
\item m2.price - price per square meter
\item surface - apartment area in square meters
\item n.rooms - number of rooms (correlated with surface)
\item district - district in which apartment is located, factor with 10 levels
\item floor - floor
\item construction.date - construction year
}
}
\keyword{apartments}
