% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_theme_drwhy.R
\name{colors_discrete_drwhy}
\alias{colors_discrete_drwhy}
\alias{colors_diverging_drwhy}
\alias{colors_breakdown_drwhy}
\title{DrWhy color palettes for ggplot objects}
\usage{
colors_discrete_drwhy(n = 2)

colors_diverging_drwhy()

colors_breakdown_drwhy()
}
\arguments{
\item{n}{number of colors for color palette}
}
\value{
color palette as vector of charactes
}
\description{
DrWhy color palettes for ggplot objects
}
