% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragons.R
\docType{data}
\name{dragons}
\alias{dragons}
\alias{dragons_test}
\title{Dragon Data}
\format{a data frame with 2000 rows and 8 columns}
\usage{
data(dragons)
}
\description{
Datasets \code{dragons} and \code{dragons_test} are artificial, generated form the same ground truth model, 
but with sometimes different data distridution.
}
\details{
Values are generated in a way to:
- have nonlinearity in year_of_birth and height
- have concept drift in the test set

\itemize{
\item year_of_birth - year in which the dragon was born. Negative year means year BC, eg: -1200 = 1201 BC
\item year_of_discovery - year in which the dragon was found.
\item height - height of the dragon in yards.
\item weight - weight of the dragon in tons.
\item scars - number of scars.
\item colour - colour of the dragon.
\item number_of_lost_teeth - number of teeth that the dragon lost.
\item life_length - life length of the dragon.
}
}
\keyword{dragons}
