% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain}
\alias{explain}
\title{Create Model Explainer}
\usage{
explain(model, ..., label = tail(class(model), 1))
}
\arguments{
\item{model}{object - a model to be explained.}

\item{...}{other parameters}

\item{label}{character - the name of the model. By default it's extracted from the 'class' attribute of the model.}
}
\value{
An object of the class 'explainer'.
It's a list with the model and additional meta-data, like model class, model name etc.
}
\description{
Black-box models may have very different structures.
This function creates a unified representation of a model, which can be further processed by various explainers.
}
\examples{

}
