% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_happiness.R
\docType{data}
\name{happiness}
\alias{happiness}
\alias{happiness_test}
\alias{happiness_train}
\title{World Happiness Report data}
\format{
two data frames with total 781 rows, 7 columns ech and rownames
}
\source{
World Happiness Report data \url{https://worldhappiness.report/}
}
\usage{
data(happiness_train)
data(happiness_test)
}
\description{
The \code{happiness_train} and \code{happiness_test} datasets are generated
based on the "World Happiness Report" at Kaggle \url{https://www.kaggle.com/datasets/unsdsn/world-happiness}.
}
\details{
It contains data for 781 countries and 7 variables. These are:
\itemize{
\item score - Happiness score
\item gdp_per_capita -  GDP per capita
\item social_support - Social support
\item healthy_life_expectancy - Healthy life expectancy
\item freedom_life_choices - Freedom to make life choices
\item generosity - Generosity
\item perceptions_of_corruption - Perceptions of corruption
}
}
\keyword{happiness_test}
\keyword{happiness_train}
