% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\docType{data}
\name{HR}
\alias{HR}
\alias{HRTest}
\alias{HR_test}
\title{Human Resources Data}
\format{a data frame with 10000 rows and 6 columns}
\usage{
data(HR)
}
\description{
Datasets \code{HR} and \code{HR_test} are artificial, generated form the same model.
Structure of the dataset is based on a real data, from Human Resources department with
information which employees were promoted, which were fired.
}
\details{
Values are generated in a way to:
- have interaction between age and gender for the 'fired' variable
- have non monotonic relation for the salary variable
- have linear effects for hours and evaluation.

\itemize{
\item gender - gender of an employee.
\item age - age of an employee in the moment of evaluation.
\item hours - average number of working hours per week.
\item evaluation - evaluation in the scale 2 (bad) - 5 (very good).
\item salary - level of salary in the scale 0 (lowest) - 5 (highest).
\item status - target variable, either `fired` or `promoted` or `ok`.
}
}
\keyword{HR}
