% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dependencies.R
\name{install_dependencies}
\alias{install_dependencies}
\title{Install all dependencies for the DALEX package}
\usage{
install_dependencies(packages = c("ingredients", "iBreakDown", "pdp",
  "ALEPlot", "breakDown", "ggpubr", "factorMerger"))
}
\arguments{
\item{packages}{which packages shall be installed?}
}
\description{
By default 'heavy' dependencies are not installed along DALEX.
This function silently install all required packages.
}
