% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_individual_diagnostics.R
\name{plot.individual_diagnostics_explainer}
\alias{plot.individual_diagnostics_explainer}
\title{Plot Instance Level Residual Diagnostics}
\usage{
\method{plot}{individual_diagnostics_explainer}(x, ...)
}
\arguments{
\item{x}{ab object with instance level residual diagnostics created with \code{\link{individual_diagnostics}} function}

\item{...}{other parameters}
}
\value{
An \code{ggplot2} object of the class \code{gg}.
}
\description{
Plot Instance Level Residual Diagnostics
}
\examples{
library("ranger")
titanic_glm_model <- ranger(survived ~ gender + age + class + fare + sibsp + parch,
                     data = titanic_imputed)
explainer_glm <- explain(titanic_glm_model,
                         data = titanic_imputed,
                         y = titanic_imputed$survived)
johny_d <- titanic_imputed[24, c("gender", "age", "class", "fare", "sibsp", "parch")]

\dontrun{
pl <- individual_diagnostics(explainer_glm, johny_d, variables = NULL)
plot(pl)

pl <- individual_diagnostics(explainer_glm, johny_d,
                       neighbors = 10,
                       variables = c("age", "fare"))
plot(pl)

pl <- individual_diagnostics(explainer_glm,
                       johny_d,
                       neighbors = 10,
                       variables = c("class", "gender"))
plot(pl)
}

}
