% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_individual_profile_explainer.R
\name{plot.individual_profile_explainer}
\alias{plot.individual_profile_explainer}
\title{Plot Variable Profile Explanations}
\usage{
\method{plot}{individual_profile_explainer}(x, ...)
}
\arguments{
\item{x}{a model to be explained, preprocessed by the \code{\link{explain}} function}

\item{...}{other parameters}
}
\value{
An object of the class \code{variable_profile_explainer}.
}
\description{
Plot Variable Profile Explanations
}
