% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_parts.R
\name{plot.model_parts}
\alias{plot.model_parts}
\title{Plot Variable Importance Explanations}
\usage{
\method{plot}{model_parts}(x, ...)
}
\arguments{
\item{x}{an object of the class \code{model_parts}}

\item{...}{other parameters described below}
}
\value{
An object of the class \code{ggplot}.
}
\description{
Plot Variable Importance Explanations
}
\section{Plot options}{


\subsection{variable_importance}{
\itemize{
 \item{max_vars}{maximal number of features to be included in the plot. default value is \code{10}}
 \item{show_boxplots}{logical if \code{TRUE} (default) boxplot will be plotted to show permutation data.}
 \item{bar_width}{width of bars. By default \code{10}}
 \item{desc_sorting}{logical. Should the bars be sorted descending? By default \code{TRUE}}
 \item{title}{the plot's title, by default \code{'Feature Importance'}}
 \item{subtitle}{a character. Plot subtitle. By default \code{NULL} - then subtitle is set to "created for the XXX, YYY model",
       where XXX, YYY are labels of given explainers.}
}
}
}

