% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_prediction.R
\name{plot.single_prediction_explainer}
\alias{plot.single_prediction_explainer}
\title{Plots Local Explanations}
\usage{
\method{plot}{single_prediction_explainer}(response1, ...)
}
\arguments{
\item{response1}{a single prediction exlainer produced with the 'single_prediction' function}

\item{...}{other explainers that shall be plotted together}
}
\value{
a ggplot2 object
}
\description{
Function 'plot.single_prediction_explainer' plots break down plots for a single prediction.
}
\examples{

}
