% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.single_variable_explainer}
\alias{plot.single_variable_explainer}
\title{Plots Marginal Responses}
\usage{
\method{plot}{single_variable_explainer}(response1, ...)
}
\arguments{
\item{response1}{a single variable exlainer produced with the 'single_variable' function}

\item{...}{other explainers that shall be plotted together}
}
\value{
a ggplot2 object
}
\description{
Function 'plot.single_variable_explainer' plots marginal responses for one or more explainers.
}
\examples{

}
