% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variable_profile.R
\name{plot.variable_profile_explainer}
\alias{plot.variable_profile_explainer}
\title{Plot Variable Profile Explanations}
\usage{
\method{plot}{variable_profile_explainer}(x, ..., geom = "aggregates")
}
\arguments{
\item{x}{a variable profile explanation, created with the \code{\link{variable_profile}} function}

\item{...}{other parameters}

\item{geom}{either \code{"aggregates"}, \code{"aggregates_profiles"}, \code{"aggregates_profiles_points"} determines which will be plotted}
}
\value{
An object of the class \code{ggplot}.
}
\description{
Plot Variable Profile Explanations
}
