% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_description.R
\name{print.description}
\alias{print.description}
\title{Print Natural Language Descriptions}
\usage{
\method{print}{description}(x, ...)
}
\arguments{
\item{x}{an individual explainer produced with the `describe()` function}

\item{...}{other arguments}
}
\description{
Generic function
}
