% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_explainer.R
\name{print.explainer}
\alias{print.explainer}
\title{Prints Explainer Summary}
\usage{
\method{print}{explainer}(explainer, ...)
}
\arguments{
\item{explainer}{a model exapliner created with the `explain` function}
}
\description{
Prints Explainer Summary
}
\examples{

}
