% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_individual_diagnostics_explainer.R
\name{print.individual_diagnostics_explainer}
\alias{print.individual_diagnostics_explainer}
\title{Print Instance Level Residual Diagnostics}
\usage{
\method{print}{individual_diagnostics_explainer}(x, ...)
}
\arguments{
\item{x}{ab object with instance level residual diagnostics created with \code{\link{individual_diagnostics}} function}

\item{...}{other parameters}
}
\description{
Generic function
}
