% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_model_diagnostics.R
\name{print.model_diagnostics}
\alias{print.model_diagnostics}
\title{Print Dataset Level Model Diagnostics}
\usage{
\method{print}{model_diagnostics}(x, ...)
}
\arguments{
\item{x}{an object with dataset level residual diagnostics created with \code{\link{model_diagnostics}} function}

\item{...}{other parameters}
}
\description{
Generic function
}
