% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_info.R
\name{print.model_info}
\alias{print.model_info}
\title{Print model_info}
\usage{
\method{print}{model_info}(x, ...)
}
\arguments{
\item{x}{- an object of class \code{model_info}}

\item{...}{- other parameters}
}
\description{
Function prints object of class \code{model_info} created with \code{\link{model_info}}
}
