% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_model_profile.R
\name{print.model_profile}
\alias{print.model_profile}
\title{Print Dataset Level Model Profile}
\usage{
\method{print}{model_profile}(x, ...)
}
\arguments{
\item{x}{an object with dataset level profile created with \code{\link{model_profile}} function}

\item{...}{other parameters}
}
\description{
Generic function
}
