% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_predict_diagnostics.R
\name{print.predict_diagnostics}
\alias{print.predict_diagnostics}
\title{Print Instance Level Residual Diagnostics}
\usage{
\method{print}{predict_diagnostics}(x, ...)
}
\arguments{
\item{x}{an object with instance level residual diagnostics created with \code{\link{predict_diagnostics}} function}

\item{...}{other parameters}
}
\description{
Generic function
}
