% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.scikitlearn_set.R
\name{print.scikitlearn_set}
\alias{print.scikitlearn_set}
\title{Prints scikitlearn_set class}
\usage{
\method{print}{scikitlearn_set}(x, ...)
}
\arguments{
\item{x}{a model created with `scikitlearn_model` function}

\item{...}{other arguments}
}
\description{
Prints scikitlearn_set class
}
\examples{
library("DALEX")
library("reticulate")
have_sklearn <- reticulate::py_module_available("sklearn.ensemble")
if(have_sklearn) {
  model <- scikitlearn_model(system.file("extdata", "gbm.pkl", package = "DALEX")) 
  print(model$params)
} else {
  print('Python testing environment is required.')
}
}
