% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_variable_profile.R
\name{print.variable_profile_explainer}
\alias{print.variable_profile_explainer}
\title{Print Dataset Level Variable Profile}
\usage{
\method{print}{variable_profile_explainer}(x, ...)
}
\arguments{
\item{x}{an object with instance level residual diagnostics created with \code{\link{variable_profile}} function}

\item{...}{other parameters}
}
\description{
Generic function
}
