% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_variable.R
\name{single_variable}
\alias{single_variable}
\title{Marginal Response for a Single Variable}
\usage{
single_variable(explainer, variable, type = "pdp", trans = I, ...)
}
\arguments{
\item{explainer}{a model to be explained, preprocessed by the 'explain' function}

\item{variable}{character - name of a single variable}

\item{type}{character - type of the response to be calculated.
Currently following options are implemented: 'pdp' for Partial Dependency and 'ale' for Accumulated Local Effects}

\item{trans}{function - a transformation/link function that shall be applied to raw model predictions}

\item{...}{}
}
\value{
An object of the class 'single_variable_explainer'.
It's a data frame with calculated average response.
}
\description{
Calculates the average model response as a function of a single selected variable.
Use the 'type' parameter to select the type of marginal response to be calculated.
Currently we have Partial Dependency and Accumulated Local Effects implemented.
}
\examples{

}
