% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_leverage.R
\name{variable_dropout}
\alias{variable_dropout}
\title{Loss from Variable Dropout}
\usage{
variable_dropout(explainer, loss_function = function(observed, predicted)
  sqrt(mean((observed - predicted)^2)), ..., type = "raw", n_sample = 1000)
}
\arguments{
\item{explainer}{a model to be explained, preprocessed by the 'explain' function}

\item{loss_function}{a function thet will be used to assess variable importance}

\item{...}{other parameters}

\item{type}{character, type of transformation that should be applied for dropout loss. 'raw' results raw drop lossess, 'ratio' returns \code{drop_loss/drop_loss_full_model} while 'difference' returns \code{drop_loss - drop_loss_full_model}}

\item{n_sample}{number of observations that should be sampled for calculations of variable leverages}
}
\value{
An object of the class 'variable_leverage_explainer'.
It's a data frame with calculated average response.
}
\description{
Loss from Variable Dropout
}
\examples{

}
