% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_effect.R
\name{variable_effect}
\alias{variable_effect}
\alias{variable_effect_partial_dependency}
\alias{variable_effect_accumulated_dependency}
\title{Dataset Level Variable Effect as Partial Dependency Profile or Accumulated Local Effects}
\usage{
variable_effect(explainer, variables, ..., type = "partial_dependency")

variable_effect_partial_dependency(explainer, variables, ...)

variable_effect_accumulated_dependency(explainer, variables, ...)
}
\arguments{
\item{explainer}{a model to be explained, preprocessed by the 'explain' function}

\item{variables}{character - name of variables to be explained}

\item{...}{other parameters}

\item{type}{character - type of the response to be calculated.
Currently following options are implemented: 'partial_dependency' for Partial Dependency and 'accumulated_dependency' for Accumulated Local Effects}
}
\value{
An object of the class 'aggregated_profiles_explainer'.
It's a data frame with calculated average response.
}
\description{
From DALEX version 1.0 this function calls the \code{\link[ingredients]{accumulated_dependency}} or
\code{\link[ingredients]{partial_dependency}} from the \code{ingredients} package.
Find information how to use this function here: \url{https://pbiecek.github.io/PM_VEE/partialDependenceProfiles.html}.
}
\examples{
HR$evaluation <- factor(HR$evaluation)

HR_glm_model <- glm(status == "fired"~., data = HR, family = "binomial")
explainer_glm <- explain(HR_glm_model, data = HR)
expl_glm <- variable_effect(explainer_glm, "age", "partial_dependency")
plot(expl_glm)

 \dontrun{
library("randomForest")
HR_rf_model <- randomForest(status == "fired" ~., data = HR)
explainer_rf  <- explain(HR_rf_model, data = HR)
expl_rf  <- variable_effect(explainer_rf, variables = "age",
                       type = "partial_dependency")
plot(expl_rf)
plot(expl_rf, expl_glm)

# Example for factor variable (with factorMerger)
expl_rf  <- variable_effect(explainer_rf, variables =  "evaluation")
plot(expl_rf)
 }

}
\references{
Predictive Models: Explore, Explain, and Debug. Human-Centered Interpretable Machine Learning \url{https://pbiecek.github.io/PM_VEE/}
}
