% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_profile.R
\name{variable_profile}
\alias{variable_profile}
\alias{variable_profile_partial_dependence}
\alias{variable_profile_accumulated_dependence}
\title{Dataset Level Variable Profile as Partial Dependency or Accumulated Local Dependency Explanations}
\usage{
variable_profile(
  explainer,
  variables = NULL,
  N = 100,
  ...,
  groups = NULL,
  k = NULL,
  center = TRUE,
  type = "partial"
)
}
\arguments{
\item{explainer}{a model to be explained, preprocessed by the 'explain' function}

\item{variables}{character - names of variables to be explained}

\item{N}{number of observations used for calculation of aggregated profiles. By default 100.}

\item{...}{other parameters that will be passed to \code{iBreakDown::break_down}}

\item{groups}{a variable name that will be used for grouping.
By default \code{NULL} which means that no groups shall be calculated}

\item{k}{number of clusters for the hclust function (for clustered profiles)}

\item{center}{shall profiles be centered before clustering}

\item{type}{the type of variable profile Either 'partial' or 'accumulated'.}
}
\value{
An object of the class 'variable_profile_explainer'.
It's a data frame with calculated average model responses.
}
\description{
This function calculates explanations on a dataset level set that explore model response as a function of selected variables.
The explanations can be calulated as Partial Dependency Profile or  Accumulated Local Dependency Profile.
Find information how to use this function here: \url{https://pbiecek.github.io/ema/partialDependenceProfiles.html}.
}
\details{
Underneath this function calls the \code{\link[ingredients]{partial_dependency}} or
\code{\link[ingredients]{accumulated_dependency}} functions from the \code{ingredients} package.
}
\examples{
titanic_glm_model <- glm(survived~., data = titanic_imputed, family = "binomial")
explainer_glm <- explain(titanic_glm_model, data = titanic_imputed)
expl_glm <- variable_profile(explainer_glm, "fare")
plot(expl_glm)

 \dontrun{
library("ranger")
titanic_ranger_model <- ranger(survived~., data = titanic_imputed, num.trees = 50,
                               probability = TRUE)
explainer_ranger  <- explain(titanic_ranger_model, data = titanic_imputed)
expl_ranger <- variable_profile(explainer_ranger)
plot(expl_ranger, geom = "aggregates_profiles")

vp_ra <- variable_profile(explainer_ranger, type = "partial", variables = c("age", "fare"))
plot(vp_ra, variables = c("age", "fare"), geom = "aggregates_profiles_points")

vp_ra <- variable_profile(explainer_ranger, type = "partial", k = 3)
plot(vp_ra, geom = "aggregates_profiles")

vp_ra <- variable_profile(explainer_ranger, type = "partial", groups = "gender")
plot(vp_ra, geom = "aggregates_profiles")

vp_ra <- variable_profile(explainer_ranger, type = "accumulated")
plot(vp_ra, geom = "aggregates_profiles")
 }

}
\references{
Explanatory Model Analysis. Explore, Explain and Examine Predictive Models. \url{https://pbiecek.github.io/ema/}
}
