% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_yhat.R
\name{yhat}
\alias{yhat}
\alias{yhat.lm}
\alias{yhat.randomForest}
\alias{yhat.svm}
\alias{yhat.gbm}
\alias{yhat.glm}
\alias{yhat.cv.glmnet}
\alias{yhat.glmnet}
\alias{yhat.ranger}
\alias{yhat.model_fit}
\alias{yhat.train}
\alias{yhat.lrm}
\alias{yhat.rpart}
\alias{yhat.function}
\alias{yhat.default}
\title{Predict Functions for various models}
\usage{
yhat(X.model, newdata, ...)

\method{yhat}{lm}(X.model, newdata, ...)

\method{yhat}{randomForest}(X.model, newdata, ...)

\method{yhat}{svm}(X.model, newdata, ...)

\method{yhat}{gbm}(X.model, newdata, ...)

\method{yhat}{glm}(X.model, newdata, ...)

\method{yhat}{cv.glmnet}(X.model, newdata, ...)

\method{yhat}{glmnet}(X.model, newdata, ...)

\method{yhat}{ranger}(X.model, newdata, ...)

\method{yhat}{model_fit}(X.model, newdata, ...)

\method{yhat}{train}(X.model, newdata, ...)

\method{yhat}{lrm}(X.model, newdata, ...)

\method{yhat}{rpart}(X.model, newdata, ...)

\method{yhat}{`function`}(X.model, newdata, ...)

\method{yhat}{default}(X.model, newdata, ...)
}
\arguments{
\item{X.model}{a model to be explained}

\item{newdata}{data.frame or matrix, observations for the prediction}

\item{...}{other parameters that will be passed to the \code{predict_function}}
}
\value{
A numeric vector of predictions, or a matrix for the `multiclass` task
}
\description{
The \code{yhat} function is a wrapper over different \code{predict_function}s for various models.
The wrapper returns a single numeric score for each new observation.
\code{yhat} uses different extraction techniques for different object classes,
e.g. for the random forest classifier, it forces the output to be probabilities for a target class,
not the classes itself.
}
\details{
Currently supported models are:
\itemize{
\item class \code{cv.glmnet} and \code{glmnet} - models created with the \pkg{glmnet} package,
\item class \code{glm} - generalized linear models created with \link[stats]{glm},
\item class \code{model_fit} - models created with the \pkg{parsnip} package,
\item class \code{lm} - linear models created with \link[stats]{lm},
\item class \code{ranger} - models created with the \pkg{ranger} package,
\item class \code{randomForest} - random forest models created with the \pkg{randomForest} package,
\item class \code{svm} - support vector machines models created with the \pkg{e1071} package,
\item class \code{train} - models created with the \pkg{caret} package,
\item class \code{gbm} - models created with the \pkg{gbm} package,
\item class \code{lrm} - models created with the \pkg{rms} package,
\item class \code{rpart} - models created with the \pkg{rpart} package,
\item classes \code{ctree} and \code{cforest} - models created with the \pkg{partykit} package.
}

More \code{predict_function}s are available in the \link[DALEXtra]{https://modeloriented.github.io/DALEXtra/} package.
}
