% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain}
\alias{explain}
\title{Create Model Explainer}
\usage{
explain(model, data = NULL, predict.function = yhat,
  label = tail(class(model), 1))
}
\arguments{
\item{model}{object - a model to be explained}

\item{data}{data.frame or marix - data that was used for fitting. If not provided then will be extracted from model fit}

\item{predict.function}{function that takes two arguments: model and new data and returns numeric vector with predictions}

\item{label}{character - the name of the model. By default it's extracted from the 'class' attribute of the model}

\item{...}{other parameters}
}
\value{
An object of the class 'explainer'.
It's a list with the model and additional meta-data, like model class, model name etc.
}
\description{
Black-box models may have very different structures.
This function creates a unified representation of a model, which can be further processed by various explainers.
}
\examples{
library("randomForest")
library("breakDown")

wine_lm_model4 <- lm(quality ~ pH + residual.sugar + sulphates + alcohol, data = wine)
wine_lm_explainer4 <- explain(wine_lm_model4, data = wine, label = "model_4v")
wine_lm_explainer4

wine_rf_model4 <- randomForest(quality ~ pH + residual.sugar + sulphates + alcohol, data = wine)
wine_rf_explainer4 <- explain(wine_rf_model4, data = wine, label = "model_rf")
wine_rf_explainer4

}
