% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fifa.R
\docType{data}
\name{fifa}
\alias{fifa}
\title{FIFA20 preprocessed data

The \code{fifa} dataset is a preprocessed \code{players_20.csv} dataset which comes as
a part of "FIFA 20 complete player dataset" at \url{Kaggle.com}.

It contains 5000 'overall' best players and 43 columns. These are:

\itemize{
\item short_name and nationality of the player (not used in modeling)
\item overall, potential, value_eur, wage_eur (4 potential target variables)
\item age, height, weight, attacking skills, defending skills, goalkeeping skills (37 variables)
}

Notable transformations:

\enumerate{
\item take 43 columns: \code{[2, 4, 6, 7, 8, 10:13, 44:77]} (R indexing)
\item take rows with \code{value_eur > 0}
\item remove duplicated rows (keep first)
\item sort rows on \code{overall} and take top \code{5000}
}


Source: \url{https://www.kaggle.com/stefanoleone992/fifa-20-complete-player-dataset}}
\format{
a data frame with 5000 rows and 43 columns
}
\source{
The \code{players_20.csv} dataset was downloaded from the \url{Kaggle.com} site and went through few transformations.
The complete dataset was obtained from
\url{https://www.kaggle.com/stefanoleone992/fifa-20-complete-player-dataset#players_20.csv} on January 1, 2020.
}
\usage{
data(fifa)
}
\description{
FIFA20 preprocessed data

The \code{fifa} dataset is a preprocessed \code{players_20.csv} dataset which comes as
a part of "FIFA 20 complete player dataset" at \url{Kaggle.com}.

It contains 5000 'overall' best players and 43 columns. These are:

\itemize{
\item short_name and nationality of the player (not used in modeling)
\item overall, potential, value_eur, wage_eur (4 potential target variables)
\item age, height, weight, attacking skills, defending skills, goalkeeping skills (37 variables)
}

Notable transformations:

\enumerate{
\item take 43 columns: \code{[2, 4, 6, 7, 8, 10:13, 44:77]} (R indexing)
\item take rows with \code{value_eur > 0}
\item remove duplicated rows (keep first)
\item sort rows on \code{overall} and take top \code{5000}
}


Source: \url{https://www.kaggle.com/stefanoleone992/fifa-20-complete-player-dataset}
}
\keyword{fifa}
