% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{loss_cross_entropy}
\alias{loss_cross_entropy}
\alias{loss_sum_of_squares}
\alias{loss_root_mean_square}
\alias{loss_accuracy}
\alias{loss_one_auc}
\title{Preimplemented Loss Functions}
\usage{
loss_cross_entropy(observed, predicted, p_min = 1e-04, na.rm = TRUE)
}
\arguments{
\item{observed}{observed scores or labels, these are supplied as explainer specific `y`}

\item{predicted}{predicted scores, either vector of matrix, these are returned from the model specific `predict_function()``}

\item{p_min}{for cross entropy, minimal value for probability to make sure that `log` will not explode}

\item{na.rm}{logical, should missing values be removed?}
}
\value{
numeric - value of the loss function
}
\description{
Preimplemented Loss Functions
}
\examples{
 \dontrun{
library("randomForest")
HR_rf_model <- randomForest(as.factor(status == "fired")~., data = HR, ntree = 100)
loss_sum_of_squares(as.numeric(HR$status == "fired"), yhat(HR_rf_model))
 }
}
