% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predict_profile.R
\name{plot.predict_profile}
\alias{plot.predict_profile}
\title{Plot Variable Profile Explanations}
\usage{
\method{plot}{predict_profile}(x, ...)
}
\arguments{
\item{x}{a model to be explained, preprocessed by the \code{\link{explain}} function}

\item{...}{other parameters}
}
\value{
An object of the class \code{variable_profile}.
}
\description{
Plot Variable Profile Explanations
}
