% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variable_profile.R
\name{plot.variable_profile_explainer}
\alias{plot.variable_profile_explainer}
\title{Plot Variable Profile Explanations}
\usage{
\method{plot}{variable_profile_explainer}(x, ..., geom = "aggregates")
}
\arguments{
\item{x}{a variable profile explanation, created with the \code{\link{variable_profile}} function}

\item{...}{other parameters}

\item{geom}{either \code{"aggregates"}, \code{"aggregates_profiles"}, \code{"aggregates_profiles_points"} determines which will be plotted}
}
\value{
An object of the class \code{ggplot}.
}
\description{
Plot Variable Profile Explanations
}
\examples{
titanic_glm_model <- glm(survived~., data = titanic_imputed, family = "binomial")
explainer_glm <- explain(titanic_glm_model, data = titanic_imputed)
expl_glm <- variable_profile(explainer_glm, "fare")
plot(expl_glm)

 \dontrun{
library("ranger")
titanic_ranger_model <- ranger(survived~., data = titanic_imputed, num.trees = 50,
                               probability = TRUE)
explainer_ranger  <- explain(titanic_ranger_model, data = titanic_imputed)
expl_ranger <- variable_profile(explainer_ranger)
plot(expl_ranger)
plot(expl_ranger, geom = "aggregates_profiles")

vp_ra <- variable_profile(explainer_ranger, type = "partial", variables = c("age", "fare"))
plot(vp_ra, variables = c("age", "fare"), geom = "aggregates_profiles_points")

vp_ra <- variable_profile(explainer_ranger, type = "partial", k = 3)
plot(vp_ra)
plot(vp_ra, geom = "aggregates_profiles")
plot(vp_ra, geom = "aggregates_profiles_points")

vp_ra <- variable_profile(explainer_ranger, type = "partial", groups = "gender")
plot(vp_ra)
plot(vp_ra, geom = "aggregates_profiles")
plot(vp_ra, geom = "aggregates_profiles_points")

vp_ra <- variable_profile(explainer_ranger, type = "accumulated")
plot(vp_ra)
plot(vp_ra, geom = "aggregates_profiles")
plot(vp_ra, geom = "aggregates_profiles_points")
 }

}
