% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.scikitlearn_model.R
\name{print.scikitlearn_model}
\alias{print.scikitlearn_model}
\title{Prints scikitlearn_model class}
\usage{
\method{print}{scikitlearn_model}(x, ...)
}
\arguments{
\item{x}{a model created with `scikitlearn_model` function}

\item{...}{other arguments}
}
\description{
Prints scikitlearn_model class
}
