% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.scikitlearn_set.R
\name{print.scikitlearn_set}
\alias{print.scikitlearn_set}
\title{Prints scikitlearn_set class}
\usage{
\method{print}{scikitlearn_set}(x, ...)
}
\arguments{
\item{x}{a model created with `scikitlearn_model` function}

\item{...}{other arguments}
}
\description{
Prints scikitlearn_set class
}
