% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_prediction.R
\name{single_prediction}
\alias{single_prediction}
\title{Explanations for a Single Prediction}
\usage{
single_prediction(explainer, observation, ...)
}
\arguments{
\item{explainer}{a model to be explained, preprocessed by the 'explain' function}

\item{observation}{a new observarvation for which predictions need to be explained}

\item{...}{other parameters}
}
\value{
An object of the class 'single_prediction_explainer'.
It's a data frame with calculated average response.
}
\description{
Explanations for a Single Prediction
}
\examples{

}
