% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_prediction.R
\name{single_prediction}
\alias{single_prediction}
\title{Explanations for a Single Prediction}
\usage{
single_prediction(explainer, observation, ...)
}
\arguments{
\item{explainer}{a model to be explained, preprocessed by the 'explain' function}

\item{observation}{a new observarvation for which predictions need to be explained}

\item{...}{other parameters}
}
\value{
An object of the class 'single_prediction_explainer'.
It's a data frame with calculated average response.
}
\description{
Explanations for a Single Prediction
}
\examples{
library("randomForest")
library("breakDown")

wine_lm_model4 <- lm(quality ~ pH + residual.sugar + sulphates + alcohol, data = wine)
wine_lm_explainer4 <- explain(wine_lm_model4, data = wine, label = "model_4v")
wine_lm_predict4 <- single_prediction(wine_lm_explainer4, observation = new.wine)
wine_lm_predict4

wine_rf_model4 <- randomForest(quality ~ pH + residual.sugar + sulphates + alcohol, data = wine)
wine_rf_explainer4 <- explain(wine_rf_model4, data = wine, label = "model_rf")
wine_rf_predict4 <- single_prediction(wine_rf_explainer4, observation = new.wine)
wine_rf_predict4

}
