% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_profile.R
\name{variable_profile}
\alias{variable_profile}
\alias{variable_profile_partial_dependence}
\alias{variable_profile_accumulated_dependence}
\title{Dataset Level Variable Profile as Partial Dependency or Accumulated Local Dependency Explanations}
\usage{
variable_profile(explainer, variables = NULL, ..., type = "partial_dependence")

variable_profile_partial_dependence(explainer, variables = NULL, ...)

variable_profile_accumulated_dependence(explainer, variables = NULL, ...)
}
\arguments{
\item{explainer}{a model to be explained, preprocessed by the 'explain' function}

\item{variables}{character - names of variables to be explained}

\item{...}{other parameters that will be passed to \code{iBreakDown::break_down}}

\item{type}{the type of variable profile Either 'partial_dependence' or 'accumulated_dependence'.}
}
\value{
An object of the class 'variable_profile_explainer'.
It's a data frame with calculated average model responses.
}
\description{
This function calculates explanations on a dataset level set that explore model response as a function of selected variables.
The explanations can be calulated as Partial Dependency Profile or  Accumulated Local Dependency Profile.
Find information how to use this function here: \url{https://pbiecek.github.io/ema/partialDependenceProfiles.html}.
}
\details{
Underneath this function calls the \code{\link[ingredients]{partial_dependency}} or
\code{\link[ingredients]{accumulated_dependency}} functions from the \code{ingredients} package.
}
\references{
Explanatory Model Analysis. Explore, Explain and Examine Predictive Models. \url{https://pbiecek.github.io/ema/}
}
